tidy up types and guards
