+++ packages/ts-client/src/reconnection.ts
@@ -1,6 +1,7 @@
 import type { Endpoint } from './webrtc';
-import type { FishjamClient, MessageEvents } from './FishjamClient';
+import type { FishjamClient } from './FishjamClient';
 import { isAuthError } from './auth';
+import { MessageEvents } from './types';
 
 export type ReconnectionStatus = 'reconnecting' | 'idle' | 'error';
